reorganize the entire locker
