+++ patch-db/Cargo.toml
@@ -12,6 +12,7 @@ version = "0.1.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [features]
+default = ["trace", "unstable"]
 debug = ["tracing"]
 trace = ["debug", "tracing-error"]
 unstable = []
