+++ patch-db/src/lib.rs
@@ -16,8 +16,6 @@ mod store;
 mod transaction;
 
 #[cfg(test)]
-mod proptest;
-#[cfg(test)]
 mod test;
 
 pub use handle::{DbHandle, PatchDbHandle};
