+++ patch-db/src/locker/proptest.rs
